Changes to Creation Page
Creation page now display data from DB
added Printhtml youtube thumbnail
changes to delete user

+++ Redir_DeleteUser.php
@@ -1,6 +1,10 @@
 <?php
 
 require_once 'autoload.php';
-Check_Admin();
-DeleteUser($_GET["ID"]);
-header("Location: MangeUsers.php");
+User::CheckLogin();
+if (isset($_GET["ID"]) && User::getSessionAccses()->hasAccsesAdmin()) {
+	DeleteUser($_GET["ID"]);
+	header("Location: MangeUsers.php");
+}
+
+
